Reformat PaymentModelTest
